bot2/Bott-tangrams-Mark--main/Bot - نسخة/requirements.txt
pyTelegramBotAPI>=4.10.0
supabase>=2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
reportlab>=3.6.0
arabic-reshaper>=0.3.0
python-bidi>=0.4.2
apscheduler>=3.11.0

pandas
openpyxl

python-dotenv

# Google AI (Gemini)
google-genai>=0.1.0

# Cloud Database & API
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.28.0
